Add checkout to testing jobs
Since the build and test jobs are processed separately -- in fact,
they can even run on different runner systems -- there is no assurance
that the test job will run against the tree produced by the build
job. (There is a dependency between them, but they are separate jobs,
and the built tree is not carried over from one to the other.)  We
need to ensure that the correct version will be used, by checking it
out.